+++ pyfredapi/series_collection.py
@@ -4,16 +4,21 @@ SeriesCollection is meant to make handling multiple series easier. Often users o
 `pyfredapi` offers the `SeriesCollection` class to streamline the process of collecting and munging the data for plotting and analysis.
 """
 
+import time
 from dataclasses import dataclass
 from typing import Callable, Dict, List, Union
 
 import pandas as pd
-import plotly.express as px
-from plotly.graph_objects import Figure
 from rich.console import Console
 
 from pyfredapi._base import _get_api_key
 from pyfredapi.series import SeriesInfo, get_series, get_series_info
+
+try:
+    import plotly.express as px
+    from plotly.graph_objects import Figure
+except ImportError:
+    pass
 
 console = Console()
 
@@ -63,23 +68,28 @@ class SeriesData:
             subtitle = f"<sup>{subtitle}</sup>"
             return f"{title}<br>{subtitle}"
 
-        fig = px.line(
-            data_frame=self.df,
-            x="date",
-            y=value_col,
-            title=format_title(
-                title=self.info.title,
-                start_date=self.info.observation_start,
-                end_date=self.info.observation_end,
-                subtitle=f"{self.info.seasonal_adjustment}, {self.info.units}",
-            ),
-            color_discrete_sequence=px.colors.qualitative.Safe,
-            labels=dict(
-                value=f"{self.info.id}",
-                date="Date",
-                variable="Series",
-            ),
-        )
+        try:
+            fig = px.line(
+                data_frame=self.df,
+                x="date",
+                y=value_col,
+                title=format_title(
+                    title=self.info.title,
+                    start_date=self.info.observation_start,
+                    end_date=self.info.observation_end,
+                    subtitle=f"{self.info.seasonal_adjustment}, {self.info.units}",
+                ),
+                color_discrete_sequence=px.colors.qualitative.Safe,
+                labels=dict(
+                    value=f"{self.info.id}",
+                    date="Date",
+                    variable="Series",
+                ),
+            )
+        except NameError as e:
+            raise ImportError(
+                "Plotly is required to create plots. You can install pyfredapi with plotly with `pip install 'pyfredapi[plot]`"
+            ) from e
 
         fig.update_layout(showlegend=False)
 
@@ -108,77 +118,64 @@ def _rename_series(
 
 
 class SeriesCollection:
-    """A collection of pyfredapi.SeriesData objects.
+    """A collection of `pyfredapi.SeriesData` objects.
 
     Useful when you need to collect and manage multiple economic series. Provides methods
     for listing metadata, collecting data, and merging data together in long or wide formats.
     """
 
-    def __init__(self, api_key: Union[str, None] = None):
-        """Create instance of SeriesCollection.
+    def __init__(
+        self,
+        series_id: Union[List[str], str, None] = None,
+        api_key: Union[str, None] = None,
+        rename: Union[Dict[str, str], Callable[[str], str], None] = None,
+        drop_realtime: bool = True,
+        sleep: float = 0.1,
+        **kwargs,
+    ):
+        """Create an instance of SeriesCollection.
 
         Parameters
         ----------
+        series_id : List[str] | str | None, optional
+            List of series IDs to add to collection.
         api_key : str | None, optional
             FRED API key. Defaults to None. If None, will search for FRED_API_KEY in environment variables.
+        drop_realtime : bool, optional
+            Indicates if you want to drop the realtime columns.
+        rename : Union[Dict[str, str], Callable[[str], str], None], optional
+            Label to give series. Defaults to series ID.
+        sleep : float, optional
+            Time to sleep between requests. Defaults to 0.1.
+        **kwargs : dict, optional
+            Additional parameters to FRED API `series/` endpoint.
+            Refer to the FRED documentation for a list of all possible parameters.
         """
+        self._data: List[SeriesData] = []
+        self.sleep = sleep
+        self.rename = rename
+        self.drop_realtime = drop_realtime
         if api_key is None:
             self.api_key = _get_api_key()
-        self.data: Dict[str, SeriesData] = {}
 
-    def rename_series(self, rename):
-        """Rename series columns."""
-        for series_id, series_data in self.data.items():
-            series_name = _rename_series(series_data=series_data, rename=rename)
+        if isinstance(series_id, str):
+            series_id = [series_id]
+        if series_id is None:
+            series_id = []
 
-            orig_col_name = [
-                c
-                for c in self.data[series_id].df.columns.tolist()
-                if c not in date_cols
-            ].pop()
-
-            self.data[series_id].df.rename(columns={orig_col_name: series_name}, inplace=True)  # type: ignore
-
-    def add_series(
-        self,
-        series_ids: Union[str, List[str]],
-        drop_realtime: bool = True,
-        rename: Union[Dict[str, str], Callable[[str], str], None] = None,
-        **kwargs,
-    ) -> None:
-        """Add series to class instance.
-
-        A request to the FRED api will be made for the series. The data will
-        be formatted as as pandas dataframe.
-
-        After adding a series, you can access it via an attribute or
-        in the responses dict.
-
-        Parameters
-        ----------
-        series_ids : Union[str, List[str]]
-            Series to add to collection.
-        drop_realtime : bool
-            Indicates if you want to drop the realtime columns.
-        rename : Union[Dict[str, str], Callable[[str], str], None]
-            Label to give series. Defaults to series ID.
-        **kwargs : dict, optional
-            Additional parameters to FRED API `series/` endpoint. Refer to the FRED documentation for a list of all possible parameters.
-        """
-        if isinstance(series_ids, str):
-            series_ids = [series_ids]
-
-        for series_id in series_ids:
-            if series_id in self.data:
-                print(f"Already have {series_id}")
+        for sid in series_id:
+            time.sleep(sleep)
+            if sid in self._data:
+                print(f"Already have {sid}")
                 continue
 
-            print(f"Requesting series {series_id}...")
+            print(f"Requesting series {sid}...")
 
-            df = get_series(series_id=series_id, api_key=self.api_key, **kwargs)
-            assert isinstance(df, pd.DataFrame)  # noqa: S101
+            series_info = get_series_info(series_id=sid)
+            series_df = get_series(series_id=sid, api_key=self.api_key, **kwargs)
+            assert isinstance(series_df, pd.DataFrame)  # noqa: S101
 
-            series_data = SeriesData(info=get_series_info(series_id=series_id), df=df)
+            series_data = SeriesData(info=series_info, df=series_df)
 
             if drop_realtime:
                 series_data.df.drop(
@@ -190,11 +187,85 @@ class SeriesCollection:
                 series_name = series_data.info.id
 
             series_data.df.rename(columns={"value": series_name}, inplace=True)
-            self.data[series_id] = series_data
+            self._data.append(series_data)
+            setattr(self, sid, series_data)
+
+    def __getitem__(self, key):
+        return [s for s in self._data if s.info.id == key].pop()
+
+    def __delitem__(self, key):
+        self.remove(key)
+
+    def __iter__(self):
+        for series_info in self._data:
+            yield series_info
+
+    def __len__(self):
+        return len(self._data)
+
+    def __rich_repr__(self):
+        yield self._data
+
+    def rename_series(self, rename):
+        """Rename series columns."""
+        for series_data in self._data:
+            series_name = _rename_series(series_data=series_data, rename=rename)
+
+            orig_col_name = [
+                c for c in series_data.df.columns.tolist() if c not in date_cols
+            ].pop()
+
+            series_data.df.rename(columns={orig_col_name: series_name}, inplace=True)
+
+    def add(self, series_ids: Union[str, List[str]], **kwargs) -> None:
+        """Add series to the collection.
+
+        A request to the FRED api will be made for the series. The data will
+        be formatted as as pandas dataframe.
+
+        After adding a series, you can access it via an attribute or
+        in the responses dict.
+
+        Parameters
+        ----------
+        series_ids : Union[str, List[str]]
+            Series to add to collection.
+        **kwargs : dict, optional
+            Additional parameters to FRED API `series/` endpoint.
+            Refer to the FRED documentation for a list of all possible parameters.
+        """
+        if isinstance(series_ids, str):
+            series_ids = [series_ids]
+
+        for series_id in series_ids:
+            time.sleep(self.sleep)
+            if series_id in self._data:
+                print(f"Already have {series_id}")
+                continue
+
+            print(f"Requesting series {series_id}...")
+
+            info = get_series_info(series_id=series_id)
+            df = get_series(series_id=series_id, api_key=self.api_key, **kwargs)
+            assert isinstance(df, pd.DataFrame)  # noqa: S101
+
+            series_data = SeriesData(info=info, df=df)
+
+            if self.drop_realtime:
+                series_data.df.drop(
+                    ["realtime_start", "realtime_end"], inplace=True, axis=1
+                )
+            if self.rename:
+                series_name = _rename_series(series_data, self.rename)
+            else:
+                series_name = series_data.info.id
+
+            series_data.df.rename(columns={"value": series_name}, inplace=True)
+            self._data.append(series_data)
             setattr(self, series_id, series_data)
 
-    def drop_series(self, series_ids: Union[str, List[str]]) -> None:
-        """Drop series from collection.
+    def remove(self, series_ids: Union[str, List[str]]) -> None:
+        """Remove series from collection.
 
         Parameters
         ----------
@@ -206,10 +277,11 @@ class SeriesCollection:
 
         for series_id in series_ids:
             try:
-                del self.data[series_id]
-            except KeyError:
-                raise ValueError from KeyError(f"No series '{series_id}' in collection")
+                series_info = [s for s in self._data if s.info.id == series_id].pop()
+            except IndexError as e:
+                raise ValueError(f"No series '{series_id}' in collection") from e
 
+            self._data.remove(series_info)
             delattr(self, series_id)
             print(f"Removed series {series_id}")
 
@@ -230,7 +302,7 @@ class SeriesCollection:
             col_name = "series"
 
         long_df_prep = []
-        for series in list(self.data.values()):
+        for series in self._data:
             series_name = [
                 c for c in series.df.columns.tolist() if c not in date_cols
             ].pop()
@@ -249,16 +321,11 @@ class SeriesCollection:
         pd.DataFrame
             Wide pandas dataframe.
         """
-        wide_df_prep = [
-            series.df.copy().set_index("date") for series in list(self.data.values())  # type: ignore
-        ]
+        wide_df_prep = [series.df.copy().set_index("date") for series in self._data]
         wide_df = pd.concat(wide_df_prep, axis=1)
         return wide_df.reset_index()
 
-    def merge_asof(
-        self,
-        base_series_id: str,
-    ) -> pd.DataFrame:
+    def merge_asof(self, base_series_id: str) -> pd.DataFrame:
         """Merge the series in the collection into a wide pandas dataframe based on nearest date.
 
         Uses pandas `merge_asof` methods to merge the data based on nearest date.
@@ -273,25 +340,29 @@ class SeriesCollection:
         pd.DataFrame
             Wide pandas dataframe.
         """
-        base_df = self.data[base_series_id].df.copy()
-        for series_data in list(self.data.values()):
+        base_series = [
+            series_info
+            for series_info in self._data
+            if series_info.info.id == base_series_id
+        ].pop()
+        base_df = base_series.df.copy()
+        for series_data in self._data:
             if base_series_id == series_data.info.id:
                 continue
             df = series_data.df.copy()
-            base_df = pd.merge_asof(left=base_df, right=df, on="date")  # type: ignore
+            base_df = pd.merge_asof(left=base_df, right=df, on="date")
 
-        return base_df  # type: ignore
+        return base_df
 
     def list_series(self) -> None:
         """List the series' id and title."""
-        for series_data in list(self.data.values()):
+        for series_data in self._data:
             console.print(f"{series_data.info.id}: {series_data.info.title}")
 
     def list_seasonality(self) -> None:
         """List the series' seasonality."""
         seasonal_adjustments = [
-            series_data.info.seasonal_adjustment
-            for series_data in list(self.data.values())
+            series_data.info.seasonal_adjustment for series_data in self._data
         ]
         distinct_seasonality = set(seasonal_adjustments)
 
@@ -301,15 +372,13 @@ class SeriesCollection:
 
         for season in distinct_seasonality:
             console.rule(f"[bold red]Series that are {season}")
-            for series_data in list(self.data.values()):
+            for series_data in self._data:
                 if series_data.info.seasonal_adjustment == season:
                     console.print(f"{series_data.info.id}: {series_data.info.title}")
 
     def list_frequency(self) -> None:
         """List the series' frequency."""
-        frequencies = [
-            series_data.info.frequency for series_data in list(self.data.values())
-        ]
+        frequencies = [series_data.info.frequency for series_data in self._data]
         distinct_freq = set(frequencies)
 
         if len(distinct_freq) == 1:
@@ -318,13 +387,13 @@ class SeriesCollection:
 
         for freq in distinct_freq:
             console.rule(f"[bold red]Series that are published {freq}")
-            for series_data in list(self.data.values()):
+            for series_data in self._data:
                 if series_data.info.frequency == freq:
                     console.print(f"{series_data.info.id}: {series_data.info.title}")
 
     def list_units(self) -> None:
         """List the series' measurement units."""
-        units = [series_data.info.units for series_data in list(self.data.values())]
+        units = [series_data.info.units for series_data in self._data]
         distinct_units = set(units)
 
         if len(distinct_units) == 1:
@@ -333,15 +402,13 @@ class SeriesCollection:
 
         for unit in distinct_units:
             console.rule(f"[bold red]Series that are measured in {unit}")
-            for series_data in list(self.data.values()):
+            for series_data in self._data:
                 if series_data.info.units == unit:
                     console.print(f"{series_data.info.id}: {series_data.info.title}")
 
     def list_end_date(self) -> None:
         """List the series' latest date."""
-        end_dates = [
-            series_data.info.observation_end for series_data in list(self.data.values())
-        ]
+        end_dates = [series_data.info.observation_end for series_data in self._data]
         distinct_end_dates = set(end_dates)
 
         if len(distinct_end_dates) == 1:
@@ -350,16 +417,13 @@ class SeriesCollection:
 
         for date in distinct_end_dates:
             console.rule(f"[bold red]Series that end on {date}")
-            for series_data in list(self.data.values()):
+            for series_data in self._data:
                 if series_data.info.observation_end == date:
                     console.print(f"{series_data.info.id}: {series_data.info.title}")
 
     def list_start_date(self) -> None:
         """List the series' earliest date."""
-        start_dates = [
-            series_data.info.observation_start
-            for series_data in list(self.data.values())
-        ]
+        start_dates = [series_data.info.observation_start for series_data in self._data]
         distinct_start_dates = set(start_dates)
 
         if len(distinct_start_dates) == 1:
@@ -368,14 +432,6 @@ class SeriesCollection:
 
         for date in distinct_start_dates:
             console.rule(f"[bold red]Series that start on {date}")
-            for series_data in list(self.data.values()):
+            for series_data in self._data:
                 if series_data.info.observation_start == date:
                     console.print(f"{series_data.info.id}: {series_data.info.title}")
-
-    def plot(self):
-        """Make a basic plotly time series plot with all series in the collection."""
-        raise NotImplementedError
-
-    def extract_series(self):
-        """Return a new instance of SeriesCollection, with a subset of data from this instance."""
-        raise NotImplementedError